refactor catalogue to componenets
