+++ components/catalogue/search-bar/search-bar.comp.tsx
