+++ modules/catalogue/item-detail/default-properties/item-properties.comp.tsx
@@ -1,8 +1,8 @@
-import ItemPropertyTitle from 'modules/shared/item-property/item-property-title.comp'
-import ItemPropertyValue from 'modules/shared/item-property/item-property-value.comp'
+import ItemPropertyTitle from 'components/item-property/item-property-title.comp'
+import ItemPropertyValue from 'components/item-property/item-property-value.comp'
 import { message } from 'i18n/src/messages'
-import { CatalogueItem } from 'types/responses'
 import React from 'react'
+import { CatalogueItem } from 'types/responses'
 
 import DisclosureComponent from '../disclosure/disclosure.comp'
