+++ modules/catalogue/item-detail/item-detail.comp.tsx
@@ -1,56 +1,38 @@
-import ProgressBarComponent from 'components/ui/progress-bar.comp'
 import { CatalogueItem } from 'types/responses'
-import { useEffect, useState } from 'react'
 
 import ItemPropertiesComponent from './default-properties/item-properties.comp'
 import ImageGalleryComponent from './gallery/image-gallery.comp'
 
 interface Props {
   images: { name: string; id: number; src: string }[]
-  item: CatalogueItem | undefined
+  item: CatalogueItem
+  groups: Array<string>
 }
 
-const ItemDetailComponent = ({ item, images }: Props) => {
-  const [groups, setGroups] = useState<Array<string>>([])
-
-  useEffect(() => {
-    if (item?.details) {
-      const uniqueDetailGroups = item.details
-        .map(item => item.propertyGroup)
-        .filter((value, index, self) => {
-          return self.indexOf(value) === index
-        })
-
-      setGroups(uniqueDetailGroups)
-    }
-  }, [item])
+const ItemDetailComponent = ({ item, images, groups }: Props) => {
   return (
     <div className="bg-white pb-10">
-      {item ? (
-        <main className="mx-auto max-w-7xl sm:px-6 sm:pt-16 lg:px-8 h-full overflow-auto">
-          <div className="mx-auto max-w-2xl lg:max-w-none">
-            <div className="lg:grid lg:grid-cols-2 lg:items-start lg:gap-x-8">
-              <ImageGalleryComponent images={images} />
+      <main className="mx-auto max-w-7xl sm:px-6 sm:pt-16 lg:px-8 h-full overflow-auto">
+        <div className="mx-auto max-w-2xl lg:max-w-none">
+          <div className="lg:grid lg:grid-cols-2 lg:items-start lg:gap-x-8">
+            <ImageGalleryComponent images={images} />
 
-              <div className="mt-10 px-4 sm:mt-16 sm:px-0 lg:mt-0">
-                <h1 className="text-3xl font-bold tracking-tight text-gray-900">{item.name}</h1>
-                <div className="mt-6">
-                  <h3 className="sr-only">Description</h3>
-                  <div
-                    className="space-y-6 text-base text-gray-700"
-                    dangerouslySetInnerHTML={{
-                      __html: item.description === '' || item.description === null ? 'No description' : item.description
-                    }}
-                  />
-                </div>
-                <ItemPropertiesComponent item={item} groups={groups} />
+            <div className="mt-10 px-4 sm:mt-16 sm:px-0 lg:mt-0">
+              <h1 className="text-3xl font-bold tracking-tight text-gray-900">{item.name}</h1>
+              <div className="mt-6">
+                <h3 className="sr-only">Description</h3>
+                <div
+                  className="space-y-6 text-base text-gray-700"
+                  dangerouslySetInnerHTML={{
+                    __html: item.description === '' || item.description === null ? 'No description' : item.description
+                  }}
+                />
               </div>
+              <ItemPropertiesComponent item={item} groups={groups} />
             </div>
           </div>
-        </main>
-      ) : (
-        <ProgressBarComponent />
-      )}
+        </div>
+      </main>
     </div>
   )
 }
